Safari-609.1.11.1/LayoutTests/js/array-sort-sparse-expected.txt

This tests that arrays and array like objects containing holes are sorted correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS testSort([,undefined,0,1]) is true
PASS testSort({length:4,1:undefined,2:0,3:1}) is true
PASS 0 in array is true
PASS 1 in array is false
PASS 0 in array is true
PASS 1 in array is false
PASS 2 in array is false
PASS successfullyParsed is true

TEST COMPLETE
